feat: move TopFilter to commons and create IAccountSettings interface
Move TopFilter sealed class from amethyst model to commons module,
and create IAccountSettings interface exposing the defaultXxxFollowList
StateFlow properties that DAL feed filters need.

This unblocks moving DAL filters to commons for the KMP iOS migration.

Changes:
- commons: Add TopFilter sealed class (with kotlinx.serialization)
- commons: Add IAccountSettings interface with 7 follow list flows
- commons: Add 'settings: IAccountSettings' to IAccount interface
- commons: Add kotlinx-serialization-json dependency
- amethyst: AccountSettings now implements IAccountSettings
- amethyst: Account.settings now overrides IAccount.settings
- amethyst: Update all TopFilter imports to commons package
